Fixed AssertionError in mixed integer logics.
LAEquality wrongly computed the type of the equality to Int even if the
value was a Real.  In this case, the variable will be wrapped by a
to_real and not be integral anymore.  However, this seems to be a good
point for optimization since equalities of the form (= i 0.5) should be
simplified to false and not be subject to the IRA-Hack.