New ArrayTheory based on DAGs.
The internal data structure for computing the weak and weak-i
equivalence classes is changed to a DAG-like structure.  We have
two outgoing edges, one for the default case (weak equivalence)
and one for the case the default case cannot be taken (weak-i
equivalence for the index of the store in the default case).
It turns out that this is enough to nicely represent the weak
equivalence and all weak-i equivalence relations for all indices.