Fixed bugs reported by Jochen.
The Clausifier did not set up the ProofTracker for flattening.  This was
changed via the API.
Added Exceptions in thrown while asserting a term to the delta debugger
friendly mode.
Fixed a bug in the delta debugger where ite terms were not correctly replaced.
This lead to an endless loop.