Fix insertVar.
Don't propagateBounds and ignore conflicts in insertVar.  Instead we
just add the variale to mPropBounds.  Additionally we check for mPropBounds
in setLiteral and checkpoint, where we can throw conflicts.

Also we don't clear the mPropBounds in pop() just in case.  Instead we
only remove the deleted variables from it.  We also made mPropBound a
TreeSet to allow quick insertion and deletion.