Centralised option handling.
Options are now handled by a centralised OptionMap instance.  This map manages
setting and retrieving options.  Furthermore it is able to reset all options to
their initial value thus fixing a bug in the old reset code.

The options are grouped into front end options and solver options.  The front
end options are specific to the ParseEnvironment and consists of
:regular-output-channel, :print-success, and :print-terms-cse.  These options
are meaningless if SMTInterpol is used through its API and thus not set up.  As
a side effect, setting the (SMTLIB-compliant) option :print-success when only
using the API will result in an UnsupportedOperationException.

The :diagnostic-output-channel has a special role.  It is neither a front end
option, nor a solver option.  It will be set up whenever SMTInterpol knows how
to change the destination of a LogProxy.  At the moment, this is the case when
the LogProxy is an instance of DefaultLogger.  In all other cases I assume the
logger will be externally configured and do not set up this option.

The remaining options are solver options.  Some of them are cached to speed up
access to their current value since they are regularly queried.