Create copy of OptionMap.
When the solver is copied, the option map is now copied as well to not destroy
the option settings in the original solver.  Exceptions to this rule are the
channel options and the verbosity option.  Since the :diagnostic-output-channel
should not be active if this feature is used, we only need to care about the
logger.  But the logger is not copied.  I don't want to add this feature since
copying a logger seems to be strange and might lead to strange effects if
multiple logger log to the same destination.  Hence, logger settings affect both
solvers in the same way.