Weakened interpolant checker.
If we received an OOM during interpolant-check-mode, we claimed the generated
interpolants were not inductive.  This claim is too strong since we only failed
to show indictivity due to OOM.  Now, we will only print a warning, but not
throw an exception stating that the generated interpolants are not inductive.