Ommit doomed tests.
The delta debugger still tested the whole right part after splitting a failure
block into a left and a right part and receiving a success on the left part.
The test on the right part cannot succeed.  Now, it will be skipped and the
right block (called a buddy) will be split on success of the left block.