Bugfixes and performance improvments.
The delta debugger did not correctly detect and remove neutral elements.
Sometimes this led to testing the current term due to errors in the removal
process.  This caused non-termination since the remove neutrals phase was always
considered successful.  I added an assertion to detect this problem.

During the whole delta debugging time, only two output reapers are used.  The
are reinitizialized on every test to reap the output of the new process streams.