Prevent models for unknown result
SMTInterpol crashed when the solver timed out but the user wanted to create a
model.  This happens, e.g., when the solver did not get to ensureIntegrals() and
thus tries to create a non-integral value for an integer variable.  This update
now tells the user that models are not available if the solver returned unknown.