Generator for tight LIRA rhombi
This generator actually takes quite some time since it computes optimal values.
The benchmarks generated by this tool come in two flavours.  For the sat
benchmarks, we allow only a few solutions.  For the unsat benchmarks, we simply
remove these solutions.  Solutions are generated based on the difference between
variable y and the next integer.  To be able to compare with yices 2, we do not
use to_int, but use a fresh integer variable.