Reworked options
All options should be available independent of the usage of the front end or the
API.  We now set up the font end options every time, but emit a warning if the
regular output channel is changed, but the front end is not activated.
The logger option was changed to simplify the option by shifting more state into
the log proxy.