Always log before execute
Reverted the patch for de{clare,fine}-{fun,sort} that first executed the command
and then logged it.  We now always first log and then execute commands.  Note
that this might turn some interaction into benchmarks with errors due to
multiple definitions.  We now defer removal of duplicated definitions to a
post-processor.