Weakened assertion in insertPropagetdLiteralBefore
This method is used when explaining unit literals to insert auxiliary
literals needed for the explanation.  Since the unit literal was
propagated the auxiliary literals must appear before the explained
literal in the stack.

Usually the literal that is explained is on the stack with the same
polarity and the assertion checked this.  However, when a literal is
set with negated polarity, although the linear solver already has the
literal in the list that it wants to propagate, the literal is already
set negated on the stack, but for simplicity still the code to explain
a unit clause is used to generate the conflict clause.  This fixes
the failed assertion.