Reworked output of rationals of sort Real
The SMTLIB standard requires us to output model values for LRA as (/ num denom)
where num and denom are numerals.  Reading in such values is only supported
through an extension that is not implemented by every solver.  Furthermore, in
LIRA logics, SMTLIB requires to use the to_real function which results in ugly
output.  With this commit I deliberately break SMTLIB-compliance for logics
involving Reals.  We now output every Real using decimals, i.e., either as a
single decimal or as the division of two decimals.  While not compatible with
other solvers, I think this is the better output in the long run.