Don't catch exceptions in checkSat.
Instead let them fall through.

The application doesn't print unknown and continue but it will abort with
an exception dump.  In that case it would be unwise to continue since we
just had an unexpected error that may lead to inconsistent data structures.
Note that SMTLibExceptions are still caught by the parser and printed as
error messages.

External applications will be able to catch the error and handle it, or
they also abort.