No hashcode in Clause.
Clause are hashed when converting the proofs.  However, we need that two
clauses with different proofs should not be considered the same.  Also
since the literals in the clause can be reordered by the DPLL engine,
the hashcode wouldn't be stable.

By removing hash code we now use object identity for equality.
Hashtables containing clauses should be changed to LinkedHashMap if they
require a deterministic order, but it looks like we currently do not
iterate over clause hash tables.

This partly fixes ultimate-pa/ultimate#352 and ultimate-pa/ultimate#353.
The regression files trigger now a completely different problem in the
interpolator, though.