Always compute exact epsilon for values
The new SOI pivoter needs the exact values anyway and variables that
violate this are considered out of bounds.  Therefore compute the exact
epsilon values for all linvars: mCurval is now a
MutableExactInfiniteNumber.

Fix a bug in CutCreator where in the LIRA case some branches are
reported as cuts.

Cleaned up the model mutator and simplified it quite a bit.