Performance fix for Cuckoo Hash
Old code had the problems that for large hashes hash1 == hash2 with high
probability leading to much more collisions and a long stash list.

We tweaked the hash code to make hash1 and hash2 independent.
Also replaced stash list by a single stash and resize when a second
stash is needed.  This may lead to very early resizes, but from some
experience, once there is a stashed element, many more will follow so it
is better to resize anyway.