Added termITEBound axioms
For an ite term of the form (ite c x (+ x 1)) we can two axioms: the ite
term must be greater equal to x and smaller equal to x + 1.  This axiom
is instantiated if all branches of an ite term only differ by a constant
and thus there is a fixed minimal and maximal element that can be
derived at pre-processing time.

Also updated the proof document (and improved it).