Changed normalization of boolean xor terms.
Previously we used = for boolean terms (if and only if).  Now we changed
it to xor.  Xor is a bit more regular than equals.  Also it helps to
avoid confusion between congruence closure and boolean reasoning.

Renamed the Utils class to LogicSimplifier, as it mainly contains the
code to simplify logical operators in the term compiler and the
clausifier.

Also the proof document was extended.