If tied, prefer variable with smallest index.
Previously the variables were just sorted by freedom and index, but the
variable whose weight let the value switch was taken.  This variable was
usually not the smallest one, which is bad for the later Bland rule
steps.

Now use a hash from freedom to FreedomLimiter and just remember the
variable with the smallest value for each distinct freedom as well as
their total weight.