BugFix: find congruent term when creating new term
If a new term is created, we need to merge it with a congruent term.
This must be the term with the lowest decision level, otherwise we would
lose the congruence on backtracking.  The previous approach to find this
term didn't work.

Instead we now go through the list of all candidates, compute for
each the decision level where they are congruent, and then choose the
term with the lowest index.