Normalize affine terms when substituting.
When substituting terms for auxiliary variables, when the auxiliary
variable occurs in an affine term and the substituted term is an affine
term or a term that equals another term in the affine sum, the resulting
term would no longer be normalized.  Since we rely in other places, that
all terms are normalized, we now detect affine terms, convert them to
SMTAffineTerm, substitute them and convert back.  This way all affine
terms stay normalized.