Bugfix: CClosure and sharing term on the fly.
If a term that is already part of a congruence class is shared, we need
to fix the mSharedTerm of all involved congruence classes.  The old code
didn't do it correctly, especially if the shared term didn't have a
shared term in "its" sub-class and there is another shared term in
the congruence class.

We also propagate LA equalities on backtrack in case it was created by
sharing a term but the corresponding CC equality was already created
earlier.  (This part is untested).