More SMT-LIB conform way to normalize rationals
Rationals normalized form is (/ (- 2) 5), instead of (/ (- 2.0) 5.0).
In LRA integer rationals are still normalized to x.0 or (- x.0), like
they are in LIRA.

Added more unit tests.

To simplify parsing, we also normalize negated numerals in LRA to a
negative BigInteger constant value.  Note that in LRA, we cannot use
Rationals to represent numeral values.  Negated numerals are allowed
to appear as numerator in rationals.

We still normalize negated decimals in LRA and LIRA to a function
application term "-" of a constant of with a positive BigDecimal value.

The unit tests should capture most of these normalizations and test
that terms that have the same written form in SMT-LIB are actually
identical.  Also parsing a syntactically correct SMT-LIB term and
converting it back to SMT-LIB should produce the same output, e.g.
"0.4" and "(/ 2 5)" are parsed to different terms.