Optimization: switch data structure from List to Set.
Non-basic integer variables were tracked as a list.
When popping levels from the solver stack,
the variables to be removed need to be searched in the list before they can be removed,
and removing from an ArrayList might also need a full traversal of the list.
Using a Set allows instant access and deletion.

Please note:
- Using a LinkedHashSet should keep the iteration order deterministic.
- For me, it is unknown whether a variable could appear several times in the list.
  The new Set only tracks "one" instance.