error message if produce-interpolants is not set
The interpolator now needs proofs and interactive-mode, which
are both enabled when produce-interpolants is set to true.  It will
still work without the produce-interpolants option, as long as
both proof and assertion tracking is enabled, but it will no longer
work if assertion tracking was not enabled.  Print a clear error
message in that case.