Quantifier Theory: extend set of interesting terms for model-based instantiation
If a variable is the lhs of an equality with a ground rhs, use all terms
for which a disequality with the rhs is set. For variable equalities,
use all pairs of terms with disequalities.

With the new option ":quant-mb-allcclasses-forvareq" set to true, all
representatives of congruence classes are used in the above cases.