Rework quantifier proofs
Now the clausifier builds the proof for a unit clause containing a
forall literal.  The clausifier also computes the quantified variables
and passes it to the quantifier theory.

Fix a problem where literals in DER proofs were not reordered correctly.
We now destruct the or proved by the DER proof and construct the or
term with the correct order.