Fix Trichotomy-Lemma.
The problem was, if the trichotomy had a LiteralReason, but that literal
could not be used for explanation, since it's on the wrong decision
level.  In that case the explanation was added in form of several
literals, which does not yield a trichotomy.

Always use a sub annotation and if that turns out to be a single literal
or subannotation, then we optimize this later.  Make sure we don't
overwrite the subannotations from Trichotomy with the one that depends
on it.