Clean up ite-Lemmas (CollectConditions).
If the ite contains the same condition (modulo negation) multiple times,
include it only once and only visit the branch with the same polarity.
Otherwise the corresponding ite lemma would not only contain the same
literal in two polarities, it would also break the proof simplifier as
it cannot guess which branch was taken.