Simplified cycle lemma.
For select edges x -> sel(x), always create the ((_ is cons) x) testers
on x itself, even if an equivalent tester already exists.  This
simplifies the lemma (and fixes a bug with proof explanation).  It also
makes interpolation simpler.  The cost is an additional CCAppTerm and a
congruence that must be propagated.