Interpolator Fix: Trivial Disequalities
Fix a problem with CC-Lemmas using "trivial" mixed disequalities like
2a != 2b + 1.  Previously, these trivial disequalities were omitted completely.
Now there are added explitly to the proof and proved separately with a
special EQ lemma having only this single literal.  This EQ instance is then
interpolated by the EQInterpolator.

Note that these are also added for disequalities like x != x + 1, where the
difference is a constant.  These can never be mixed and are therefore always
interpolated as either true or false.