Bugfix CCEquality that looks like LA.
CCEqualities are distinguished from LAEqualities by not having a
zero on the right-hand-side.  Previously the CCProofGenerator would move any
constant on the left-hand-side, but that doesn't work anymore for
trivial disequalities like 0=1.  So we check explicitly for 0.

The other occurrence in CClosure.createCCEquality is left unchanged, as it
is never used for trivial disequalities.