Don't unshare ccterm when deconstructing
Instead of unsharing CCTerms when they are removed on pop, they are only
unshared by the clausifier, explicitly.  If the term is removed the shared term
will not be used anyways.
There was a bug before where pop() happens in different order and the same
term was unshared twice.

Fixes #145